requirements.txt
youtube-dl==2021.4.26
python-dotenv==0.19.2